Change fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
 int main(void)
 {
 	int i;
-	long long arr[50];
+	float arr[50];
 
 	arr[0] = 1;
 	arr[1] = 2;
@@ -17,7 +17,7 @@ int main(void)
 	}
 	for (i = 0; i < 50; i++)
 	{
-		printf("%lld", arr[i]);
+		printf("%.0f", arr[i]);
 		if (i < 49)
 		{
 			printf(", ");
